server: add external-logical-cluster-admin flag
The current logical-cluster-admin config is used for both internal
shard-direct access, and also for access via the front-proxy.

This won't work when the CA certs for the shard and front-proxy
are different, so we add a new flag to specify an external config
that enables access via the front-proxy - the existing config should
be configured to enable access direct to the shards.

This is derived from 339a071a which was reverted in 3854f1bc

Co-Authored-By: Dr. Stefan Schimanski <stefan.schimanski@gmail.com>